Refactor import paths for trainer and predictor modules, remove deprecated model data file, and enhance API documentation for predictor and trainer requests
