+++ local_test/get_clean_images.py
@@ -1,4 +1,4 @@
-from src.api.integrations.integrations import Integrations
+from src.entities.machine_learning.src.integrations import Integrations
 
 if __name__ == "__main__":
     integrations = Integrations()
