+++ local_test/get_extract_area_at_coordinates.py
@@ -1,4 +1,4 @@
-from src.api.integrations.integrations import Integrations
+from src.entities.machine_learning.src.integrations import Integrations
 
 if __name__ == "__main__":
     lon = 0.0
