+++ local_test/get_visualize_image_coordinates.py
@@ -1,6 +1,6 @@
 
 import os
-from src.api.integrations.integrations import Integrations
+from src.entities.machine_learning.src.integrations import Integrations
 
 if __name__ == "__main__":
     input_dir = os.path.join(os.getcwd(), "downloads", "extracted_area")
