+++ src/api/trainer/docs.py
