+++ src/entities/machine_learning/src/copernicus/constants.py
