+++ src/entities/machine_learning/src/copernicus/copernicus.py
